requirements.txt
aiohttp==3.7.4.post0
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
base58==2.1.0
beautifulsoup4==4.10.0
bitarray==1.2.2
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.6
cytoolz==0.11.0
dj-database-url==0.5.0
Django==3.2.8
django-filter==21.1
django-heroku==0.3.1
djangorestframework==3.12.4
eth-abi==2.1.1
eth-account==0.5.5
eth-hash==0.3.2
eth-keyfile==0.5.1
eth-keys==0.3.3
eth-rlp==0.2.1
eth-typing==2.2.2
eth-utils==1.10.0
gunicorn==20.1.0
hexbytes==0.2.2
idna==3.2
ipfshttpclient==0.8.0a2
jsonschema==3.2.0
lru-dict==1.1.7
Markdown==3.3.4
multiaddr==0.0.9
multidict==5.1.0
netaddr==0.8.0
parsimonious==0.8.1
protobuf==3.18.0
psycopg2==2.9.1
pycryptodome==3.10.1
pyrsistent==0.18.0
pytz==2021.3
requests==2.26.0
rlp==2.0.1
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.2
toolz==0.11.1
typing-extensions==3.10.0.2
urllib3==1.26.6
varint==1.0.2
web3==5.23.1
websockets==9.1
whitenoise==5.3.0
yarl==1.6.3
